requirements.txt
torch
torchvision
torchmetrics
scikit-learn
numpy
tensorflow
tqdm
Pillow
